massive refactoring

+++ src/Models/Object.php
@@ -13,6 +13,7 @@ use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Config;
 use Carbon\Carbon;
 use Magnetar\Tariffs\Presenters\ValidatePresenter;
+use Magnetar\Tariffs\References\ObjectReference;
 use Magnetar\Tariffs\References\UserBalanceReference;
 use Magnetar\Tariffs\Services\UserBalanceService;
 use DB;
@@ -59,13 +60,14 @@ class Object extends Model {
                 continue;
 
             $check_settings = $module_check->settings;
-
             if(!isset($check_settings['count']) || (isset($check_settings['count']) && $check_settings['count'] == 0)) {
 
-                $user_module_check = UserObject::where('user_id', $user_id)->where('module_id', $module_key)->first();
+                if($this->type_id == ObjectReference::getTypeId(ObjectReference::MAGNETAR_TARIFFS_TARIFFS)) { // todo::
+                    $user_module_check = UserObject::where('user_id', $user_id)->where('module_id', $module_key)->first();
 
-                if(isset($user_module_check))
-                    $user_module_check->delete();
+                    if (isset($user_module_check))
+                        $user_module_check->delete();
+                }
 
                 $ar_ids = [];
                 if ($module_check->group > 0) {
@@ -79,27 +81,24 @@ class Object extends Model {
                     if ($user_biggest_module_check > 0)
                         continue;
                 }
-
             }
-// TODO: start, удаляем дубли, возможно нужно смотреть по expired_at
-            $module_check = Module::select('magnetar_tariffs_modules.settings')
-                ->leftJoin('magnetar_tariffs_user_objects', 'magnetar_tariffs_modules.id', '=', 'magnetar_tariffs_user_objects.module_id')
-                ->where('magnetar_tariffs_user_objects.user_id', $user_id)
-                ->where('magnetar_tariffs_user_objects.module_id', $module_key)
-                ->first();
-// TODO: end
-            if(isset($module_check)) {
-                $check_settings = $module_check->settings;
-                if(!isset($check_settings['count']) || (isset($check_settings['count']) && $check_settings['count'] == 0))
-                    continue;
-            }
+//// TODO: start, удаляем дубли, возможно нужно смотреть по expired_at
+//            $module_check = Module::select(Module::TABLE_NAME.'.settings')
+//                ->leftJoin(UserObject::TABLE_NAME, Module::TABLE_NAME.'.id', '=', UserObject::TABLE_NAME.'.module_id')
+//                ->where(UserObject::TABLE_NAME.'.user_id', $user_id)
+//                ->where(UserObject::TABLE_NAME.'.module_id', $module_key)
+//                ->first();
+//// TODO: end
+//            if(isset($module_check)) {
+//                $check_settings = $module_check->settings;
+//                if(!isset($check_settings['count']) || (isset($check_settings['count']) && $check_settings['count'] == 0))
+//                    continue;
+//            }
 
             $user_tariff = new UserObject();
 
             if(isset($data['price']) && is_array($data['price'])) {
-
                 $user_tariff->price = current($data['price'])['price'];
-
                 foreach ($data['price'] as $interval => $price) {
                     $date = new Carbon();
                     $date_check = $date->add(new \DateInterval($interval));
@@ -108,15 +107,11 @@ class Object extends Model {
                         $user_tariff->price = $price['price'];
                     else
                         break;
-
                 }
-
             }
 
             $insert_data = ['active' => true];
-
             $user_tariff->user_id = $user_id;
-
             $module = Module::find($module_key);
 
             if(!$module)
@@ -127,26 +122,23 @@ class Object extends Model {
             if(!isset($setting_arr['count']))
                 $user_tariff->object_id = $this->id;
 
-            if(isset($setting_arr['count']) && $setting_arr['count'] == 1 && isset($data['count'])) {
-
+            if(isset($setting_arr['count']) && $setting_arr['count'] == 1 && isset($data['count']))
                 $insert_data['count'] = $data['count'];
-
-                if($data['refresh_period']) {
-
-                    $insert_data['refresh_period'] = $data['refresh_period'];
-                    $date = new Carbon();
-                    $insert_data['refresh_in'] = $date->add(new \DateInterval($data['refresh_period']))->toIso8601String();
-
-                }
-
-            } else if(isset($setting_arr['count']) && $setting_arr['count'] == 1)
+            else if(isset($setting_arr['count']) && $setting_arr['count'] == 1)
                 continue;
 
-            if(isset($data['base_price'])) {
+            if(isset($data['refresh_period']) && $data['refresh_period']) {
+                $insert_data['refresh_period'] = $data['refresh_period'];
+                $date = new Carbon();
+                $insert_data['refresh_in'] = $date->add(new \DateInterval($data['refresh_period']))->toIso8601String();
+            }
 
+            if(isset($data['customs']))
+                $insert_data['customs'] = $data['customs'];
+
+            if(isset($data['base_price'])) {
                 $insert_data['base_price'] = $data['base_price'];
                 UserBalanceService::create($user_id, UserBalanceReference::BUY, $data['base_price'], ['name' => $module->name]);
-
             }
 
             $user_tariff->module_id = $module->id;
@@ -159,12 +151,9 @@ class Object extends Model {
 
             if(isset($user_tariff->price) && $user_tariff->price > 0)
                 $daily_price += $user_tariff->price;
-
         }
 
         if($daily_price > 0)
             UserBalanceService::create($user_id, UserBalanceReference::DAILY_BUY, $daily_price);
-
     }
-
 }
